Added files via upload
The files attached are different parts of the box drawn using Fusion 360 as well as the code for Boxzy. Although the Boxzy did not cut the pieces correctly according to the codes due to certain unknown reasons, the drawings are a good reference for future replications and production of the Collapsable Box. 

Files:
-f3d - Fusion 360 files
-nc - Coded files